update filenames, reduce circular dependencies
