+++ src/plugins/drawing-tool/model/drawing-content.ts
@@ -1,17 +1,18 @@
 import { types, Instance } from "mobx-state-tree";
 import { exportDrawingTileSpec } from "./drawing-export";
 import { importDrawingTileSpec, isDrawingTileImportSpec } from "./drawing-import";
-import { DrawingObjectDataType } from "./drawing-objects";
 import { StampModel, StampModelType } from "./stamp";
 import { ITileExportOptions, IDefaultContentOptions } from "../../../models/tools/tool-content-info";
 import { ToolMetadataModel, ToolContentModel } from "../../../models/tools/tool-types";
 import { safeJsonParse } from "../../../utilities/js-utils";
 import { Logger, LogEventName } from "../../../lib/logger";
 import {
-  DefaultToolbarSettings, DrawingToolChange, DrawingToolCreateChange, DrawingToolDeleteChange, DrawingToolMoveChange,
-  DrawingToolUpdate, DrawingToolUpdateChange, kDrawingToolID, ToolbarModalButton, ToolbarSettings
+  DrawingToolChange, DrawingToolCreateChange, DrawingToolDeleteChange, DrawingToolMoveChange,
+  DrawingToolUpdate, DrawingToolUpdateChange, kDrawingToolID, ToolbarModalButton
 } from "./drawing-types";
 import { ImageObjectSnapshot } from "../objects/image";
+import { DrawingObjectSnapshotUnion } from "../components/drawing-object-manager";
+import { DefaultToolbarSettings, ToolbarSettings } from "./drawing-basic-types";
 
 
 export const computeStrokeDashArray = (type?: string, strokeWidth?: string|number) => {
@@ -218,7 +219,7 @@ export const DrawingContentModel = ToolContentModel
             const change = safeJsonParse<DrawingToolChange>(changeJson);
             switch (change?.action) {
               case "create": {
-                const createData = change.data as DrawingObjectDataType;
+                const createData = change.data as DrawingObjectSnapshotUnion;
                 if (createData.type !== "image") {
                   break;
                 }
